create general skeleton of the website
Add folder "components", and files "TheHeader.vue" and "TheFooter.vue" inside it that
are used in the default layout. Put in TheHeader some of the landmarks that we have to put (links
to the new pages)

Add folder "layouts", and the file "default.vue" layout that is showed in every
page

Add pages "allAreas", "allProjects" and "contacts" as examples of some
pages that we have to create.

Delete something because that was just a useless example.